+++ core/domain/src/commonMain/kotlin/io/github/kabos/data/TsudanumaTimetable.kt
@@ -1,12 +1,12 @@
 package io.github.kabos.data
 
-import io.github.kabos.BusRouteName
-import io.github.kabos.StationName
-import io.github.kabos.TimetableRow
-import io.github.kabos.Tsu0506
-import io.github.kabos.Tsu07
-import io.github.kabos.Tsu08
-import io.github.kabos.WeekTimetable
+import io.github.kabos.bus.core.model.BusRouteName
+import io.github.kabos.bus.core.model.StationName
+import io.github.kabos.bus.core.model.TimetableRow
+import io.github.kabos.bus.core.model.Tsu0506
+import io.github.kabos.bus.core.model.Tsu07
+import io.github.kabos.bus.core.model.Tsu08
+import io.github.kabos.bus.core.model.WeekTimetable
 
 // https://www.navitime.co.jp/diagram/bus/00043845/00032183/1/
 internal val tsudanumaTsu0506 = WeekTimetable(
