+++ core/domain/src/commonMain/kotlin/io/github/kabos/repository/DefaultTimetableRepository.kt
@@ -3,12 +3,12 @@ package io.github.kabos.repository
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
-import io.github.kabos.BusRouteName
-import io.github.kabos.StationName
-import io.github.kabos.Tsu0506
-import io.github.kabos.Tsu07
-import io.github.kabos.Tsu08
-import io.github.kabos.WeekTimetable
+import io.github.kabos.bus.core.model.BusRouteName
+import io.github.kabos.bus.core.model.StationName
+import io.github.kabos.bus.core.model.Tsu0506
+import io.github.kabos.bus.core.model.Tsu07
+import io.github.kabos.bus.core.model.Tsu08
+import io.github.kabos.bus.core.model.WeekTimetable
 import io.github.kabos.data.takinoiTsu07
 import io.github.kabos.data.takinoiTsu08
 import io.github.kabos.data.tsudanumaTsu0506
