+++ core/domain/src/commonMain/kotlin/io/github/kabos/repository/DummyTimetableRepository.kt
@@ -2,9 +2,9 @@ package io.github.kabos.repository
 
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
-import io.github.kabos.BusRouteName
-import io.github.kabos.StationName
-import io.github.kabos.WeekTimetable
+import io.github.kabos.bus.core.model.BusRouteName
+import io.github.kabos.bus.core.model.StationName
+import io.github.kabos.bus.core.model.WeekTimetable
 
 /**
  * Debug Repository
