+++ core/domain/src/commonTest/kotlin/io/github/kabos/GetBusDepartureTimeUseCaseTest.kt
@@ -2,6 +2,12 @@ package io.github.kabos
 
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
+import io.github.kabos.bus.core.model.BusRouteName
+import io.github.kabos.bus.core.model.DayType
+import io.github.kabos.bus.core.model.StationName
+import io.github.kabos.bus.core.model.TimetableCell
+import io.github.kabos.bus.core.model.TimetableRow
+import io.github.kabos.bus.core.model.WeekTimetable
 import io.github.kabos.repository.DummyTimetableRepository
 import kotlinx.datetime.LocalTime
 import kotlin.test.BeforeTest
