+++ core/model/src/commonMain/kotlin/io/github/kabos/bus/core/model/FixedClock.kt
@@ -1,4 +1,4 @@
-package io.github.kabos
+package io.github.kabos.bus.core.model
 
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
