+++ core/model/src/commonTest/kotlin/io/github/kabos/bus/core/model/DayTypeTest.kt
@@ -1,4 +1,4 @@
-package io.github.kabos
+package io.github.kabos.bus.core.model
 
 import kotlinx.datetime.Instant
 import kotlin.test.Test
