+++ feature/clock/src/commonMain/kotlin/io/github/kabos/StationSelectDialog.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
+import io.github.kabos.bus.core.model.StationName
 
 @Composable
 fun StationSelectDialog(
