+++ feature/clock/src/commonMain/kotlin/io/github/kabos/TimelineSection.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.unit.sp
 import com.github.michaelbull.result.getOr
 import com.github.michaelbull.result.map
 import io.github.kabos.ClockContract.UiAction
+import io.github.kabos.bus.core.model.TimetableCell
 import io.github.kabos.extension.subtract
 import io.github.kabos.extension.toHHmm
 import io.github.kabos.extension.tommss
